requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlmodel==0.0.14
sqlalchemy==2.0.23
sqladmin==0.16.1
instaloader==4.10.3
openai==1.3.7
python-dotenv==1.0.0
pillow==10.1.0
aiofiles==23.2.1
httpx==0.25.2
pydantic==2.5.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0